refactor: deduplicate timeAgo, formatStaked, and StatusBadge
- Move timeAgo() and formatStaked() to shared app/lib/helpers.ts
- Extract StatusBadge into app/components/StatusBadge.tsx with all
  status variants (finalized, justified, pending, proposed, missed)
- Update home-client, epoch-detail, and epochs-list to import from
  shared locations
