+++ ExplorerFrontend/app/epochs/[query]/epochs-client.tsx
@@ -6,8 +6,9 @@ import axios from 'axios';
 import config from '../../../config';
 import { useQuery } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
-import { formatNumberWithCommas } from '../../lib/helpers';
+import { formatNumberWithCommas, timeAgo, formatStaked } from '../../lib/helpers';
 import SearchBar from '../../components/SearchBar';
+import StatusBadge from '../../components/StatusBadge';
 
 // ── Types ────────────────────────────────────────────────────────────────────
 
@@ -30,44 +31,6 @@ interface EpochsData {
 // ── Helpers ──────────────────────────────────────────────────────────────────
 
 const ITEMS_PER_PAGE = 15;
-
-function timeAgo(unixSeconds: number): string {
-  const now = Math.floor(Date.now() / 1000);
-  const diff = now - unixSeconds;
-  if (diff < 0) return 'just now';
-  if (diff < 60) return `${diff}s ago`;
-  if (diff < 3600) return `${Math.floor(diff / 60)}m ago`;
-  if (diff < 86400) return `${Math.floor(diff / 3600)}h ago`;
-  return `${Math.floor(diff / 86400)}d ago`;
-}
-
-function formatStaked(gwei: string): string {
-  try {
-    const val = BigInt(gwei || '0');
-    const qrlBase = val / BigInt(1_000_000_000);
-    const remainder = val % BigInt(1_000_000_000);
-    const decimalStr = remainder.toString().padStart(9, '0').replace(/0+$/, '');
-    const result = formatNumberWithCommas(qrlBase.toString());
-    return decimalStr.length > 0 ? `${result}.${decimalStr} QRL` : `${result} QRL`;
-  } catch {
-    return '0 QRL';
-  }
-}
-
-// ── Status Badge ─────────────────────────────────────────────────────────────
-
-function StatusBadge({ status }: { status: string }) {
-  const styles: Record<string, string> = {
-    finalized: 'bg-emerald-500/15 text-emerald-400 border-emerald-500/20',
-    justified: 'bg-blue-500/15 text-blue-400 border-blue-500/20',
-    pending: 'bg-yellow-500/15 text-yellow-400 border-yellow-500/20',
-  };
-  return (
-    <span className={`inline-flex items-center px-2 py-0.5 rounded text-[11px] font-medium border ${styles[status] || styles.pending}`}>
-      {status.charAt(0).toUpperCase() + status.slice(1)}
-    </span>
-  );
-}
 
 // ── Data Fetching ────────────────────────────────────────────────────────────
